Make workspace-tabs-store a ports-and-adapters unit test
